refactor: move files to their respective folders
