+++ include/Serialization/Json/IJsonValue.h
@@ -1,8 +1,9 @@
 // Copyright 2021 the Drakhtar authors. All rights reserved. MIT license.
 
 #include <string>
+#include "Serialization/IValue.h"
 
-class IValue {
+class IJsonValue : public IValue {
  protected:
   [[nodiscard]] inline static std::string serialize(
       const std::string& value) noexcept {
